add --yes to example workflow
Version 2.0 of cosign is asking users to confirm that information is
stored in public transparency logs. This interactive interaction does
not work in automated CI pipelines like GitHub actions and breaks
builds that are now starting to use 2.0. Adding --yes removes the
required user interaction and enables automated builds again.

Signed-off-by: Sebastian Hoß <seb@xn--ho-hia.de>